:star2: feat: move backend extension points to main.js and store/index.js
