+++ src/backend/firebase/store/blips.js
