+++ src/store/index.js
@@ -2,12 +2,21 @@ import Vue from 'vue'
 import VueX from 'vuex'
 import users from './user'
 import blips from './blips'
+import appConfig from '../config'
+import fireback from '../backend/firebase/index'
 
 Vue.use(VueX)
 
+let backend
+switch (appConfig.backend.type.toLocaleLowerCase()) {
+  case 'firebase':
+    backend = fireback
+    break
+}
+
 export const store = new VueX.Store({
   modules: {
-    users,
-    blips
+    users: users(backend),
+    blips: blips(backend)
   }
 })
